Flowrate div 0 fix

+++ ulc_mm_package/QtGUI/acquisition.py
@@ -16,7 +16,7 @@ from ulc_mm_package.QtGUI.gui_constants import ACQUISITION_PERIOD
 
 class Acquisition(QObject):
     update_liveview = pyqtSignal(np.ndarray)
-    update_scopeop = pyqtSignal(np.ndarray, int)
+    update_scopeop = pyqtSignal(np.ndarray, float)
 
     def __init__(self):
         super().__init__()
